Social/requirements.txt
chromadb>=0.5.0
sentence-transformers>=2.2.2
pyyaml>=6.0
markdown-it-py>=3.0.0
beautifulsoup4>=4.12.0
